sdpa fwd (frost): SM100 wide templates re-resolve the per-batch lengths at the persistent handoff
prefill_d256_fp8, prefill_d256_mxfp8 and prefill_d512_mxfp8 take the next
tile's KV-loop bounds from the scheduler payload and resolved eff_seqlen_kv /
eff_seqlen_q only before the persistent loop, so after a handoff to a tile of
another batch the softmax's bottom-right diagonal and the new keyless-row test
kept the previous batch's lengths (codex). Both are re-resolved right after the
payload's batch_idx read in every warp role that uses them; the KV-length
helper accepts the raw arith batch index the correction role reads there.

Multi-wave regression (192 Q tiles on a 148-SM part, three batches of
different lengths, bottom-right): red before this change on d512 mxfp8
(max|O-ref| 1.91) and on d256 fp8 (2.10 -- that one predates the PR: its
per-element diagonal already used the stale length), green after.

+++ python/cudnn/sdpa/fwd/kernels/_common_blackwell.py
@@ -707,7 +707,7 @@ def make_sdpa_helpers(
     def _resolve_seqlen_kv(seq_kv_lens_tensor, batch_idx, scalar_seqlen_kv):
         if cutlass.const_expr(CFG.SEQ_KV_LENS_PRESENT == 1):
             arr = cutlass.make_array_view(seq_kv_lens_tensor)
-            return cutlass.Int32(arr[batch_idx])
+            return cutlass.Int32(arr[cutlass.Int32(batch_idx)])  # batch_idx may be a raw arith value after a payload read
         return scalar_seqlen_kv
 
     @cute.jit
